p1/proj_env/requirements.txt
selenium
requests
beautifulsoup4
lxml
pywebview